multi_agent/langchain/requirements.txt
# LangChain and Azure AI Inference
langchain>=0.3.27
langchain-azure-ai>=0.1.5
langchain-core>=0.3.72
langchain-community>=0.3.27
langchain-openai>=0.3.28

# Azure AI Inference SDK
azure-ai-inference>=1.0.0b4
azure-core>=1.29.0
azure-identity>=1.15.0

# MCP and async support
mcp>=1.0.0
asyncio-mqtt>=0.13.0
aiofiles>=23.0.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.8.0
